backend/requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib

pydantic
pydantic-settings
sqlalchemy
alembic

google-auth
oauth2client
python-dotenv
email-validator

fastapi[all]
authlib
uvicorn
psycopg2-binary
python-decouple
itsdangerous
python-dateutil
parse
beautifulsoup4


python-jose[cryptography]
cryptography